update install script

+++ bootstrap.sh
@@ -3,4 +3,14 @@
 # A bootstrap script to setup the .dotfiles folder.
 # This file is licensed under the Do What The Fuck You Want To Public License.
 
-ln -s ./nvim ~/.config/nvim
+read -n1 -p "Would you like to install the nvim config (will overwrite)? [Y/n]" isntnvim
+case $instnvim in
+    y|Y)
+        echo "Deleting config..."
+        rm -r ~/.config/nvim 
+        ln -s ./nvim ~/.config/nvim
+        ;;
+    *) # Take everything else to be a no
+        echo "Skipping nvim config..."
+        ;;
+esac
